fix(timestamp_when_filtering): moved macro to pragma-entities + used
