+++ pragma-entities/src/lib.rs
@@ -2,8 +2,11 @@ pub mod connection;
 pub mod db;
 pub mod dto;
 pub mod error;
+pub mod macros;
 pub mod models;
 pub mod schema;
+
+pub use models::entries::entry_error::EntryError;
 
 // exporting for idiomatic use
 pub use error::{adapt_infra_error, InfraError};
@@ -11,7 +14,7 @@ pub use models::{
     checkpoint_error::CheckpointError,
     currency::Currency,
     entry::{Entry, NewEntry},
-    entry_error::{EntryError, VolatilityError},
+    entry_error::VolatilityError,
     future_entry::{FutureEntry, NewFutureEntry},
     publisher::{NewPublisher, Publishers},
     publisher_error::PublisherError,
